This update adds "repair failure", representing a case where repair fails, independent of end misrejoining, as seen in low dose-rate, low-LET exposures.
This is build to reflect observed events fit in the analytic Medras model. Breaks which are correctly joined now have a chance of undergoing 'repair failure', where a misrepair event occurs with another previously undamaged piece of DNA.

Misrepair code now takes in the distribution of chromosome sizes, which it uses to sample where the other end of the misrepair occurs. As SDD files do not provide spatial information on where chromosomes are located, this is done probabilistically to reflect fit rates for X-ray exposures from the analytic Medras.

Misrepair events are first sampled to be inter- or intra-chromosome to give the correct rate for exchanges and dicentrics. For inter-chromosome exchanges, position on the partner chromosome is sampled randomly. For intra-chromosome exchanges, positions are first sampled to either be a small or large deletion, again to reflect the overall rate from previous fitting. Small deletions are sampled within the large deletion range (default 3 MBP) from the initial break site. Large deletions are sampled randomly from the rest of the chromosome.

These changes also required updates to the repair kinetics code, which is now stand-alone, to ensure DSB kinetics are appropriately reflected.

Minor other updates elsewhere in code to pass through and process relevant data.

This feature can be modified by changing the "repairFailure" variable in medrasrepair.py